requirements.txt
django
mysqlclient
django-cors-headers
djangorestframework
drf_yasg==1.21.7
pillow